Refactor python coverage
