+++ utbot-python/src/main/kotlin/org/utbot/python/PythonTestGenerationConfig.kt
@@ -2,8 +2,8 @@ package org.utbot.python
 
 import org.utbot.framework.codegen.domain.RuntimeExceptionTestsBehaviour
 import org.utbot.framework.codegen.domain.TestFramework
-import org.utbot.python.evaluation.coverage.CoverageOutputFormat
-import org.utbot.python.evaluation.coverage.PythonCoverageMode
+import org.utbot.python.coverage.CoverageOutputFormat
+import org.utbot.python.coverage.PythonCoverageMode
 import java.nio.file.Path
 
 data class TestFileInformation(
